requirements.txt
pydantic
pytest
requests
python-dotenv